requirement.txt
Django==4.1.3
mysqlclient
djangorestframework
pyjwt
djangorestframework-simplejwt
django-rest-swagger